requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
python-decouple>=3.8
pydantic[email]>=2.5.0
razorpay>=1.4.0
twilio>=8.10.0
redis>=5.0.0
celery>=5.3.0
pillow>=10.0.0
aiofiles>=23.0.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
